condensed general functions into file reading and date conversions
